requirements.txt
ipython==8.17.2
pydantic==4.8.0
pyyaml==6.0.1
python-dotenv==1.0.0
pyTelegramBotAPI==4.14.0